Make mso detection work similar to what file/file does (#587)
* Make mso detection work similar to what file/file does

https://github.com/file/file/blob/7c62d696b06e53fc5be015c41a57513278ac6c54/magic/Magdir/msooxml
The algorithms is not 100% percent reliable. For example, a
zero compression zip containing a docx will still sometimes be detected
as docx instead of zip (it depends on how many files and the order of
files in the zip)

Second thing in this PR is removing some test data fixtures.
From now, I'll try as much as possible to write regular unit tests
without relying on test file fixtures. https://github.com/gabriel-vasile/mimetype/issues/575#issuecomment-2347642574
related #550 #575
closes #400

* zipContains: remove unnecessary zip sig check

The check is already done in parent function.